Moved tests alongside code
This aligns the browser package with the other packages.
